+++ CHANGES.md
@@ -24,6 +24,7 @@ Changes from ns-3.42 to ns-3-dev
 
 ### Changes to existing API
 
+* (core) Merged files `unix-fd-reader.cc` and `win32-fd-reader.cc` into a new unified file `fd-reader.cc`.
 * (lr-wpan) Attribute `macBeaconPayload` in `MacPibAttributes` is now a std::vector<uint8_t> instead of a packet pointer.
 * (lr-wpan) Removes the word `address` from the MAC address prefix when `LOG_PREFIX_FUNC` is used.
 * (lr-wpan) Removes the word `address` from the CSMA-CA logs prefix when `LOG_PREFIX_FUNC` is used.
@@ -49,6 +50,7 @@ Changes from ns-3.41 to ns-3.42
 -------------------------------
 
 ### New API
+
 * (antenna) Added `CircularApertureAntennaModel` class which characterizes the antenna gain pattern of the reflector antenna with circular aperture described in 3GPP TR 38.811 v15.4.0, Section 6.4.1
 * (core) Objects now can be aggregated to multiple objects though the `Object::UnidirectionalAggregateObject` function. Objects aggregated in such a way can not use `GetObject` to access the objects they are aggregated to.
 * (core) Added `TestVector` iterators and dot product operator for `Vector2D` and `Vector3D` types